+++ CHANGELOG.md
@@ -8,7 +8,8 @@
  - Fixed OpenGL ES 3.2 not working.
  - Add support for `IntVec{2|3|4}` and `UnsignedIntVec{2|3|4}` uniform types.
  - Add a work-around for Radeon drivers crashing with 32+ texture units.
- - Add support for boolean uniforms.
+ - Add support for `boolean` uniforms.
+ - Add support for `BoolVec{2|3|4}` uniform types.
 
 ## Version 0.8.5 (2015-08-12)